+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - `:ref-fillvector` constructs vector of N copies of top :ref item, using a second scalar to determine relative size of N (few, some, many, lots)
 - `:ref-cyclevector` constructs vector of N items by cycling through :ref stack, using a second scalar to determine relative size of N (few, some, many, lots)
 - added new `:vectorized` instructions to `:vector` as appropriate
+- changed `:x-print` instruction to return the string result of `pr-str`
 
 ## 0.1.21